refactor(wasm): extract drain_and_log_csg_diagnostics into api/csg_diagnostics.rs
The 270-LOC drain_and_log_csg_diagnostics console formatter has zero IfcAPI
state coupling. Move it verbatim out of api/mod.rs into a sibling
api/csg_diagnostics.rs, leaving the IfcAPI struct and every other item in
place. A private `use csg_diagnostics::drain_and_log_csg_diagnostics;` in
mod.rs keeps gpu_meshes.rs's `super::drain_and_log_csg_diagnostics` call
resolving unchanged; the moved fn reaches the still-private `set_js_prop`
via `use super::set_js_prop`, so its body is byte-identical.

Pure code motion (0 lines changed). Warning set identical to baseline; cargo
test -p ifc-lite-wasm green (12); wasm32 build clean.
